+++ src/super_gradients/examples/quantization/resnet_qat_example.py
@@ -5,17 +5,17 @@ from torch import nn
 import super_gradients
 from super_gradients import Trainer
 from super_gradients.common.object_names import Models
-from super_gradients.modules.quantization.resnet_bottleneck import QuantBottleneck as sg_QuantizedBottleneck
 from super_gradients.training import MultiGPUMode
 from super_gradients.training import models as sg_models
 from super_gradients.training.dataloaders import imagenet_train, imagenet_val
 from super_gradients.training.metrics import Accuracy, Top5
 from super_gradients.training.models.classification_models.resnet import Bottleneck
 from super_gradients.training.models.classification_models.resnet import Bottleneck as sg_Bottleneck
-from super_gradients.training.utils.quantization.calibrator import QuantizationCalibrator
-from super_gradients.training.utils.quantization.core import QuantizedMetadata
-from super_gradients.training.utils.quantization.export import export_quantized_module_to_onnx
-from super_gradients.training.utils.quantization.selective_quantization_utils import SelectiveQuantizer
+from super_gradients.training.utils.quantization import QuantBottleneck as sg_QuantizedBottleneck
+from super_gradients.training.utils.quantization import QuantizationCalibrator
+from super_gradients.training.utils.quantization import QuantizedMetadata
+from super_gradients.training.utils.quantization import export_quantized_module_to_onnx
+from super_gradients.training.utils.quantization import SelectiveQuantizer
 
 
 def naive_quantize(model: nn.Module):
