+++ src/super_gradients/modules/__init__.py
@@ -13,7 +13,6 @@ from super_gradients.modules.skip_connections import (
     BackboneInternalSkipConnection,
     HeadInternalSkipConnection,
 )
-from super_gradients.common.abstractions.abstract_logger import get_logger
 from super_gradients.common.registry.registry import ALL_DETECTION_MODULES
 
 from super_gradients.modules.base_modules import BaseDetectionModule
@@ -61,40 +60,3 @@ __all__ = [
     "SSDHead",
     "SupportsReplaceNumClasses",
 ]
-
-logger = get_logger(__name__)
-try:
-    # flake8 respects only the first occurence of __all__ defined in the module's root
-    from .quantization import QuantBottleneck  # noqa: F401
-
-    from .quantization import QuantResidual  # noqa: F401
-    from .quantization import QuantSkipConnection  # noqa: F401
-    from .quantization import QuantCrossModelSkipConnection  # noqa: F401
-    from .quantization import QuantBackboneInternalSkipConnection  # noqa: F401
-    from .quantization import QuantHeadInternalSkipConnection  # noqa: F401
-
-    from .quantization import QuantSTDCBlock  # noqa: F401
-    from .quantization import QuantAttentionRefinementModule  # noqa: F401
-    from .quantization import QuantFeatureFusionModule  # noqa: F401
-    from .quantization import QuantContextPath  # noqa: F401
-
-    quant_extensions = [
-        "QuantBottleneck",
-        "QuantResidual",
-        "QuantSkipConnection",
-        "QuantCrossModelSkipConnection",
-        "QuantBackboneInternalSkipConnection",
-        "QuantHeadInternalSkipConnection",
-        "QuantSTDCBlock",
-        "QuantAttentionRefinementModule",
-        "QuantFeatureFusionModule",
-        "QuantContextPath",
-    ]
-
-except (ImportError, NameError, ModuleNotFoundError) as import_err:
-    logger.debug(f"Failed to import pytorch_quantization: {import_err}")
-    quant_extensions = None
-
-
-if quant_extensions is not None:
-    __all__.extend(quant_extensions)
